add Tag and author model and controller + tag view
Created a controller and model for author and added a has many
relationship in the author model and the post model to associate them. I
created a controller and a model for tag and added a HABTM relationship in
the model of tag and post to associate them. I made a add view and the
method for adding tags. Updated post add to also include tags for a post
from a menu and then set the tags to be displayed in the post view.